refactor: resolved merge conflicts
