+++ src/main/java/com/chat/yourway/dto/common/EmailMessageInfoDto.java
@@ -1,9 +1,11 @@
 package com.chat.yourway.dto.common;
 
-import com.chat.yourway.model.EmailMessageType;
+import com.chat.yourway.model.email.EmailMessageType;
 
 public record EmailMessageInfoDto(String username,
                                   String email,
                                   String uuidToken,
                                   String path,
-                                  EmailMessageType emailMessageType) {}
+                                  EmailMessageType emailMessageType) {
+
+}
