+++ src/main/java/com/chat/yourway/dto/response/ApiErrorResponseDto.java
@@ -1,4 +1,4 @@
-package com.chat.yourway.exception;
+package com.chat.yourway.dto.response;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -6,7 +6,7 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @RequiredArgsConstructor
-public class ApiErrorResponse {
+public class ApiErrorResponseDto {
 
   private final HttpStatus httpStatus;
   private final String message;
