--- src/main/java/com/chat/yourway/dto/response/ErrorResponseDto.java
@@ -1,11 +1,0 @@
-package com.chat.yourway.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ErrorResponseDto {
-    private int status;
-    private String message;
-}
