+++ src/main/java/com/chat/yourway/exception/EmailSendingException.java
@@ -1,12 +1,6 @@
 package com.chat.yourway.exception;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
 public class EmailSendingException extends RuntimeException {
-
   public EmailSendingException(String message) {
     super(message);
   }
