+++ src/main/java/com/chat/yourway/exception/InvalidCredentialsException.java
@@ -1,12 +1,6 @@
 package com.chat.yourway.exception;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
 public class InvalidCredentialsException extends RuntimeException {
-
   public InvalidCredentialsException(String message) {
     super(message);
   }
