+++ src/main/java/com/chat/yourway/exception/InvalidTokenException.java
@@ -1,10 +1,5 @@
 package com.chat.yourway.exception;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
 public class InvalidTokenException extends RuntimeException {
 
   public InvalidTokenException(String message) {
