+++ src/main/java/com/chat/yourway/exception/TokenNotFoundException.java
@@ -1,0 +1,7 @@
+package com.chat.yourway.exception;
+
+public class TokenNotFoundException extends RuntimeException {
+    public TokenNotFoundException(String message) {
+        super(message);
+    }
+}
