+++ src/main/java/com/chat/yourway/service/EmailSenderService.java
@@ -1,7 +1,7 @@
 package com.chat.yourway.service;
 
+import com.chat.yourway.dto.common.EmailMessageDto;
 import com.chat.yourway.exception.EmailSendingException;
-import com.chat.yourway.model.email.EmailSend;
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
 import lombok.RequiredArgsConstructor;
@@ -22,7 +22,7 @@ public class EmailSenderService {
   private final JavaMailSender javaMailSender;
 
   @Async
-  public void sendEmail(EmailSend request) {
+  public void sendEmail(EmailMessageDto request) {
     if (request == null) {
       throw new IllegalArgumentException("EmailSend request cannot be null");
     }
