+++ src/test/java/com/chat/yourway/integration/extension/PostgresExtension.java
@@ -9,19 +9,20 @@ import org.testcontainers.utility.DockerImageName;
 
 public class PostgresExtension implements Extension, BeforeAllCallback, AfterAllCallback {
 
-    private static final PostgreSQLContainer<?> container = new PostgreSQLContainer<>(DockerImageName.parse("postgres:15.3"))
-            .withEnv("TZ", "UTC");
+  private static final PostgreSQLContainer<?> container = new PostgreSQLContainer<>(
+      DockerImageName.parse("postgres:14.6"))
+      .withEnv("TZ", "UTC");
 
-    @Override
-    public void beforeAll(ExtensionContext extensionContext) {
-        container.start();
-        System.setProperty("JDBC_URL", container.getJdbcUrl());
-        System.setProperty("POSTGRES_USERNAME", container.getUsername());
-        System.setProperty("POSTGRES_PASSWORD", container.getPassword());
-    }
+  @Override
+  public void beforeAll(ExtensionContext extensionContext) {
+    container.start();
+    System.setProperty("JDBC_URL", container.getJdbcUrl());
+    System.setProperty("POSTGRES_USERNAME", container.getUsername());
+    System.setProperty("POSTGRES_PASSWORD", container.getPassword());
+  }
 
-    @Override
-    public void afterAll(ExtensionContext extensionContext) {
-        //should do nothing, testcontainers will shut down the container
-    }
+  @Override
+  public void afterAll(ExtensionContext extensionContext) {
+    //should do nothing, testcontainers will shut down the container
+  }
 }
