+++ src/test/resources/application.properties
@@ -22,8 +22,9 @@ spring.data.redis.password=${REDIS_PASSWORD}
 logging.level.org.springframework=ERROR
 logging.level.com.example=DEBUG
 
-# WebSocket config:
+#WebSocket config:
 socket.dest-prefixes=/topic, /specific
+socket.specific-prefix=/specific
 socket.app-prefix=/app
 socket.endpoint=/chat
 
@@ -34,6 +35,12 @@ security.jwt.secret-key=jwt-secret-key
 security.jwt.expiration=3600000
 # 7 days
 security.jwt.refresh-token.expiration=604800000
+
+#CORS:
+cors.allowed-origins=http://localhost:8080
+cors.allowed-methods=GET, POST, PUT
+cors.allowed-headers=Authorization
+cors.allow-credentials=true
 
 # SMTP Configuration
 spring.mail.host=smtp.office365.com
